Create diy-part1.sh

+++ devices/ramips/mt7621/diy-part1.sh
@@ -15,6 +15,11 @@
 
 
 # JDC-1配置
-cp -rf $GITHUB_WORKSPACE/patch/JDC-1.tar.gz $GITHUB_WORKSPACE/openwrt
-tar -zxvf JDC-1.tar.gz
-rm -rf JDC-1.tar.gz
+cat .config | while read content
+do
+if [ $content = jdcloud_re-sp-01b=y];then
+  cp -rf $GITHUB_WORKSPACE/patch/JDC-1.tar.gz $GITHUB_WORKSPACE/openwrt
+  tar -zxvf JDC-1.tar.gz
+  rm -rf JDC-1.tar.gz
+fi
+done
